requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-filter>=24.1
drf-yasg>=1.21.7
django-cors-headers>=4.2.0
python-decouple>=3.8
psycopg2-binary>=2.9.9   # Only required if using PostgreSQL
